backend/requirements.txt
Django==2.2.2
djangorestframework==3.9.4
pytz==2019.1
sqlparse==0.3.0
pytest-django==3.5.0
